Create sepeare component from Details Popup
