+++ client/src/Components/index.ts
@@ -2,3 +2,4 @@ export { FormInput } from "./FormInput/FormInput";
 export { FormTextArea } from "./FormTextArea/FormTextArea";
 export { LogEntryForm } from "./LogEntryForm/LogEntryForm";
 export { Rating } from "./Rating/Rating";
+export { DetailsPopup } from "./DetailsPopup/DetailsPopup";
